lab2/requirements.txt
strands-agents
boto3
python-dotenv
bedrock-agentcore
bedrock-agentcore-starter-toolkit
